Add Clase BaseABM
