+++ UI.Web/BaseABM.aspx
@@ -1,0 +1,15 @@
+﻿<%@ Page Language="C#" AutoEventWireup="true" CodeBehind="BaseABM.aspx.cs" Inherits="UI.Web.BaseABM" %>
+
+<!DOCTYPE html>
+
+<html xmlns="http://www.w3.org/1999/xhtml">
+<head runat="server">
+    <title></title>
+</head>
+<body>
+    <form id="form1" runat="server">
+        <div>
+        </div>
+    </form>
+</body>
+</html>
